Refactor cookie.jsx: divide hooks to tokenAction.jsx, useToken.jsx
