+++ src/hooks/tokenAction.jsx
@@ -1,0 +1,16 @@
+import { useSetRecoilState } from 'recoil';
+import { accessTokenState, getAccessToken } from './useToken'; // 경로는 실제 파일 위치에 맞게 조정하세요.
+import { useRecoilValue } from 'recoil';
+import { useCookie } from './Cookie';
+
+export function ReissueAccessToken() {
+    const { getCookies } = useCookie();
+    const setAccessToken = useSetRecoilState(accessTokenState);
+    const cookies = getCookies();
+    const refreshToken = cookies.refreshToken;
+    const userId = cookies.userId; // userId를 얻는 로직이 필요한 경우 여기에 추가
+    // getAccessTokenSelector와 같은 로직 필요 - 현재 예제에는 해당 내용이 명시되어 있지 않음
+    // 해당 부분에 대한 구현이 필요합니다.
+    const newAccessToken = 'your_new_access_token'; // 임시 토큰 값, 실제 로직으로 대체 필요
+    setAccessToken(newAccessToken);
+}
